--- a/src/index.js
+++ b/src/index.js
@@ -1,5 +1,6 @@
 /* eslint-disable no-unused-vars */
 
+import './scss/tailwind.scss';
 import './scss/index.scss';
 
 // Make sure JS is enabled and the window is ready.