--- a/src/scss/tailwind.scss
+++ b/src/scss/tailwind.scss
@@ -1,0 +1,6 @@
+//-----------------------------------------
+// Tailwind Files
+//-----------------------------------------
+@tailwind base;
+@tailwind components;
+@tailwind utilities;