@gregrickaby 2021-02-10
For Next.js projects we've had to do the same thing. 

`index.css` is _only_ for importing the default Tailwind directives.

Everything custom goes into `global.css`